rag-service/requirements.txt
# RAG Service Dependencies
# Install with: pip install -r requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Vector Search & Embeddings
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2

# PDF Text Extraction
PyPDF2==3.0.1

# Document Processing
python-docx==1.1.0
tiktoken==0.5.2

# AI Generation (choose one or both)
openai==1.12.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
numpy>=1.26.0,<2.0
